Inventory will look for and record failure to access folder - possible to retrive the actual file with Invoke-CMLootHunt Improved download handling 	Won't download duplicates 	Put files in directories by extension Saved file name is <folder: filext>\<first 4 chars of hash>_<original filename> Invoke-CMLootExtract for bulk extraction .MSI files Better feedback / logging

+++ CMLoot.ps1
@@ -30,6 +30,12 @@ Enumerates files available on SCCM share and saves it to a file, excludes extens
 
     )
     
+    # Print help if no commands
+    if($PSBoundParameters.Count -eq 0) {
+        Get-Help $MyInvocation.MyCommand -Detailed
+        return
+    }
+
     # Add .INI to Exclusions
     if ($ExcludeExtension) {
         $ExcludeList = @()
@@ -38,6 +44,19 @@ Enumerates files available on SCCM share and saves it to a file, excludes extens
         }
     }
     
+    $OutfileNoAccess = $OutFile.Split(".")[0] + "_noaccess." + $OutFile.Split(".")[1]
+
+    # Check if OutFile exists.
+    if (Test-Path -Path $OutFile -PathType Leaf) {
+        Write-Output "[*] OutFile exists."
+        Remove-Item -Path $OutFile -Confirm
+    }
+
+    if (Test-Path -Path $OutfileNoAccess -PathType Leaf) {
+        Write-Output "[*] OutFile exists."
+        Remove-Item -Path $OutfileNoAccess -Confirm
+    }
+
     # Grab all shares on host and fetch all .ini-files SCCMContentLib\DataLib share
     (net view $SCCMHost /all) | % {
         if($_.IndexOf(' Disk ') -gt 0){
@@ -47,24 +66,46 @@ Enumerates files available on SCCM share and saves it to a file, excludes extens
 
                 # Folder fetching
                 $folders = Get-ChildItem -Path ('\\' + $SCCMHost + '\' + $share + '\DataLib\') -Directory -ErrorAction SilentlyContinue
-                $files = For($i = 0; $i -lt $folders.count; $i++) {
+                For($i = 0; $i -lt $folders.count; $i++) {
+                    
                     # Update progress bar for current folder
                     Write-Progress -Activity "Scanning files." -CurrentOperation ('Collecting INI Files in: '+ $folders[$i].FullName) -PercentComplete (($i+1) / $folders.count * 100) -Status ("Folder {0} of {1}" -f ($i + 1), $folders.count)
                 
                     # INI file fetching
-                    Get-ChildItem -Path $folders[$i].FullName -File -Recurse -Exclude $ExludeList -Include *.INI |
+                    Get-ChildItem -Path $folders[$i].FullName -File -Recurse -Exclude $ExludeList -Include *.INI -ErrorAction SilentlyContinue -ErrorVariable +fileErr |
                         ForEach-Object -Process {
-                        if ($_.PSIsContainer) {
-                            continue
-                        }
-                        if ($ExcludeList) {
-                            if (($_ -notmatch ($ExcludeList -join '|'))) {
+                            if ($_.PSIsContainer) {
+                                    continue
+                                }
+
+                            if ($ExcludeList) {
+                                if (($_ -notmatch ($ExcludeList -join '|'))) {
+                                    $_.FullName.Substring(0, $_.FullName.Length-4) | Add-Content -Path $OutFile
+                                }
+                            } else {
                                 $_.FullName.Substring(0, $_.FullName.Length-4) | Add-Content -Path $OutFile
                             }
-                        } else {
-                            $_.FullName.Substring(0, $_.FullName.Length-4) | Add-Content -Path $OutFile
                         }
-                    }
+                }
+
+            # Check errors if access is denied = interesting file
+            foreach($ex in $fileErr)
+                {
+                    if ($ex.Exception -is [System.UnauthorizedAccessException]) 
+                        {
+                            Write-Output ("[*] Access to {0} is denied, adding to {1}" -f $ex.TargetObject, $OutfileNoAccess) ""
+                            $ex.TargetObject.Split("\")[-1] | Add-Content -Path $OutfileNoAccess
+                            
+                        }
+                }
+            # Write help text
+            if (Test-Path -Path $OutFile -PathType Leaf)
+                {
+                    Write-Output ("[*] Inventory created. Use Invoke-CMLootDownload -InventoryFile {0} to download files. See Get-Help Invoke-CMLootDownload for options." -f $Outfile) "" 
+                }
+            if (Test-Path -Path $OutfileNoAccess -PathType Leaf)
+                {
+                    Write-Output ("[*] Inaccessible DataLib content detected. Use Invoke-CMLootHunt -SCCMHost {0} -NoAccessFile {1} to search for correlating files." -f $SCCMHost, $OutfileNoAccess) ""
                 }
             }
         }
@@ -88,6 +129,7 @@ Author: Tomas Rzepka (@1njected) / WithSecure
 Invoke-CMLootDownload -OutFolder CMLootOut -InventoryFile .\sccmfiles.txt -GridSelect
 
 Opens previously collected inventory file and opens GridView to select files to download.
+WARNING: Opening large inventory file might exhaust RAM.
 
 
 .EXAMPLE 
@@ -113,6 +155,11 @@ Opens previously collected inventory file and downloads all files with .ps1 exte
         [Parameter(Mandatory=$false)] [switch]$GridSelect
     )
     
+    # Print help if no commands
+    if($PSBoundParameters.Count -eq 0) {
+        Get-Help $MyInvocation.MyCommand -Detailed
+        return
+    }
     
     # Lookup file location, format FileLib\<first 4 chars of hash\<full file hash>
     function getfilelocation([string] $file) {
@@ -122,16 +169,28 @@ Opens previously collected inventory file and downloads all files with .ps1 exte
 
     }
 
-    
     function downloadfile($File) {
-        $FileSource = getfilelocation($File)
-        $DestFile = $OutFolder + "\" + $FileSource.Split("\")[$_.Lenght-1] + "_" + $File.Split("\")[$_.Lenght-1]
-        Copy-Item $FileSource -Destination $DestFile
+        # Output stucture <folder: ext>\<first 4 chars of hash>_<original filename>
+        $subfolder = $OutFolder + "\" + $File.Split(".")[-1]
+        
+        if (-not (Test-Path -Path $subfolder -PathType Container))
+            {
+                New-Item $subfolder -ItemType Directory -ErrorAction Ignore| Out-Null
+            }
 
+        $FileSource = getfilelocation($File)
+        $DestFile = $subfolder + "\" + $FileSource.Split("\")[-1].Substring(0,4) + "_" + $File.Split("\")[-1]
+        
+        if (-not (Test-Path -Path $DestFile -PathType Leaf))
+            {
+                Write-Output ("Downloading {0}" -f $File)
+                Copy-Item $FileSource -Destination $DestFile
+            } else {
+                Write-Output ("Already downloaded, skipping {0}" -f $File)
+            }
     }
 
-    
-    # Create output dir
+    # Create output base dir
     New-Item $OutFolder -ItemType Directory -ErrorAction Ignore| Out-Null
     
     # Perform action depending on user parameters
@@ -150,5 +209,125 @@ Opens previously collected inventory file and downloads all files with .ps1 exte
     ElseIf ($Extension) {
         Get-Content -Path $InventoryFile | Select-String -Pattern $Extension$ | ForEach-Object { $files += @($_.ToString()) }
         foreach ($item in $files) { downloadfile($item) }
+    }
+}
+
+
+function Invoke-CMLootHunt {
+    <#
+    
+    .SYNOPSIS
+    
+    Locate the actual files for which the DataLib content is denied.
+    
+    Author: Tomas Rzepka (@1njected) / WithSecure
+    
+    
+    .EXAMPLE 
+    
+    Invoke-CMLootHunt -SCCMHost sccm01.domain.local -NoAccessFile sccmfiles.txt.noaccess
+    
+    Try to match ACL denies from inventory and try to locate the actual file.
+    
+    #>
+        [CmdletBinding()]
+        param(
+            [Parameter(Mandatory=$true)] [string]$SCCMHost,
+            [Parameter(Mandatory=$true)] [string]$NoAccessFile,
+            [Parameter(Mandatory=$false)] [string]$OutFolder = "CMLootOut"
+        )
+        
+    # Print help if no commands
+    if($PSBoundParameters.Count -eq 0) {
+        Get-Help $MyInvocation.MyCommand -Detailed
+        return
+    }
+
+    # Create output base dir
+    New-Item $OutFolder -ItemType Directory -ErrorAction Ignore| Out-Null
+    
+    # Build array of lines in file
+    foreach($line in Get-Content $NoAccessFile) { $ACLSearch += $line }
+
+    # Find SCCMContentLib share on SCCM host, check all .INI files in FileLib and match content.
+    (net view $SCCMHost /all) | % {
+        if($_.IndexOf(' Disk ') -gt 0){
+            $share = $_.Split('      ')[0]
+
+            if ($share -match 'SCCMContentLib') {
+
+                # Folder fetching
+                $folders = Get-ChildItem -Path ('\\' + $SCCMHost + '\' + $share + '\FileLib\') -Directory -ErrorAction SilentlyContinue
+                For($i = 0; $i -lt $folders.count; $i++) {
+                    
+                    # Update progress bar for current folder
+                    Write-Progress -Activity "Scanning files." -CurrentOperation ('Collecting INI Files in: '+ $folders[$i].FullName) -PercentComplete (($i+1) / $folders.count * 100) -Status ("Folder {0} of {1}" -f ($i + 1), $folders.count)
+                
+                    # INI file fetching
+                    Get-ChildItem -Path $folders[$i].FullName -File -Recurse -Exclude $ExludeList -Include *.INI -ErrorAction SilentlyContinue -ErrorVariable +fileErr |
+                        ForEach-Object -Process {
+                            if ($_.PSIsContainer) {
+                                    continue
+                                }
+                            # Extract reference to DataLib
+                            $content = (Get-Content -Path $_.FullName | Select -Skip 1) -split "="
+
+                            # Match NoAccess against content
+                            if ($ACLSearch.contains($content[0])) 
+                                {
+                                    $sourcefile =  $_.DirectoryName + "\" + $_.FullName.Split("\")[-1].Split(".")[0]
+
+                                    Write-Output ("[*] File {0} match {1}. Copying to outfolder." -f $sourcefile, $content[0]) ""
+                                    Copy-Item $sourcefile -Destination $OutFolder
+                                }
+                        }
+                }
+            }
+        }
+    }
+}
+
+function Invoke-CMLootExtract {
+    <#
+    
+    .SYNOPSIS
+    
+    Extract file(s). For now, just .MSI files.
+    
+    Author: Tomas Rzepka (@1njected) / WithSecure
+    
+    
+    .EXAMPLE 
+    
+    Invoke-CMLootExtract -Path C:\Temp\CMLootOut\msi
+    
+    Extract MSI files, create a folder in same directory where the extracted files are located.
+    
+    #>
+        [CmdletBinding()]
+        param(
+            [Parameter(Mandatory=$true)] [string]$Path
+        )
+        
+    # Print help if no commands
+    if($PSBoundParameters.Count -eq 0) {
+        Get-Help $MyInvocation.MyCommand -Detailed
+        return
+    }
+
+    # Look for files without extension (files with hash as filename) and try to extract
+    Get-ChildItem -Path $Path -File -Filter "*." |
+        ForEach-Object -Process {
+            Write-Output ("Processing {0}..." -f $_.FullName)
+            $targetdir = $_.DirectoryName + "\_" + $_.BaseName
+            Start-Process "msiexec" -ArgumentList "/a $($_.FullName) /qb TARGETDIR=$targetdir" -Wait
+        }
+    
+    # Look for .msi extension and try to extract
+    Get-ChildItem -Path $Path -File -Filter "*.msi" |
+        ForEach-Object -Process {
+            Write-Output ("Processing {0}..." -f $_.FullName)
+            $targetdir = $_.DirectoryName + "\_" + $_.BaseName
+            Start-Process "msiexec" -ArgumentList "/a $($_.FullName) /qb TARGETDIR=$targetdir" -Wait
     }
 }
